EDA_CAR_SALES
"This dataset contains data for more than 9.5K cars sale in Ukraine. Most of them are used cars so it opens the possibility to analyze features related to car operation. This is a subset of all car data in Ukraine. Using this We will analyze the various parameters of used car sales in Ukraine." ---- Problem Statement

Conclusion:
Sedan cars sold maximum.
Price is increasing as the engine value is increasing.
The price and mileage goes down as engine values decreasing.